Attempt to use nix
Tests aren't building on this fork, so we need take it out here.
Since Cardano is using this we at least have some confidence in its
correctness.

If at anytime we update this dependency based on some changes by IOHK
folks (rather than rewriting this ourselves in ocaml), we should
re-evaluate doing the changes on this branch